requirements.txt
requests
beautifulsoup4
lxml
python-telegram-bot
python-dotenv